refactor(server): /api/projects 라우트 분할 (Plan C #C9 3단계)
- src/server/routes/projects.ts 신규: 7개 라우트 캡슐화 (CRUD + error-state + pause/resume)
- validateAndNormalizePath 헬퍼를 projects 전용 모듈로 이동 (dashboard-api에서 제거)
- safeError 적용 (10여 회 sanitizeErrorMessage(getErrorMessage) 치환)
- dashboard-api.ts에서 8개 unused import 정리 (ProjectConfig/normalize/isPathSafe/
  validateConfig/addProjectToConfig 등)
- dashboard-api.ts 1858 → 1593줄 (-265)

남은 도메인: jobs/config/setup/stats/repositories/health/doctor/sse
